+++ src/entries/dm-embed.ts
@@ -12,7 +12,7 @@ declare global {
 
 const init = async () => {
 
-	// Wait `.dm-player` to be ready first before do everything
+	// Wait `.dm-player` to be ready first before get the element
 	await waitFor( () => document.querySelectorAll('.dm-player').length > 0, 500, 2000, "Timeout to get DM placeholder");
 	const el: NodeListOf<HTMLDivElement> = document.querySelectorAll('.dm-player');