+++ src/player/interfaces/infVideo.ts
@@ -4,4 +4,6 @@ export default interface InfVideo {
     title?: string;
     description?: string;
     'owner.avatar_190_url'?: string;
+    thumbnail_240_url?: string;
+    duration?: number;
 }