+++ src/video/video-manager.ts
@@ -1,0 +1,14 @@
+// TODO: Separate video part from player
+
+export default class VideoManager {
+
+    constructor() {
+
+    }
+
+    render() {
+
+    }
+
+
+}